+++ c/column.h
@@ -15,6 +15,7 @@
 #include "memorybuf.h"
 #include "types.h"
 #include "stats.h"
+#include "python/list.h"
 
 class DataTable;
 class BoolColumn;
@@ -84,7 +85,7 @@ public:
   static Column* open_mmap_column(SType, int64_t nrows, const std::string& filename);
   static Column* new_xbuf_column(SType, int64_t nrows, Py_buffer* pybuffer);
   static Column* new_mbuf_column(SType, MemoryBuffer*, MemoryBuffer*);
-  static Column* from_pylist(PyObject* list, int stype0 = 0, int ltype0 = 0);
+  static Column* from_pylist(PyyList& list, int stype0 = 0, int ltype0 = 0);
 
   Column(const Column&) = delete;
   Column(Column&&) = delete;
@@ -95,9 +96,10 @@ public:
   virtual size_t elemsize() const = 0;
   virtual bool is_fixedwidth() const = 0;
 
-  inline void* data() const { return mbuf->get(); }
-  inline void* data_at(size_t i) const { return mbuf->at(i); }
-  inline const RowIndex& rowindex() const { return ri; }
+  void replace_rowindex(const RowIndex& newri);
+  void* data() const { return mbuf->get(); }
+  void* data_at(size_t i) const { return mbuf->at(i); }
+  const RowIndex& rowindex() const { return ri; }
   size_t alloc_size() const;
   virtual int64_t data_nrows() const = 0;
   PyObject* mbuf_repr() const;