+++ c/datatable.h
@@ -62,6 +62,7 @@ class DataTable {
     ~DataTable();
     DataTable* delete_columns(int*, int);
     void apply_na_mask(DataTable* mask);
+    void replace_rowindex(const RowIndex& newri);
     void reify();
     DataTable* rbind(DataTable**, int**, int, int64_t);
     DataTable* cbind(DataTable**, int);